# yukka-ui/src/openApi/openApi.json
{
  "openapi": "3.0.1",
  "info": {
    "title": "Spring Security OpenApi",
    "description": "Dokumentacja OpenApi dla Spring Security",
    "termsOfService": "Terms of service",
    "contact": {
      "name": "Łukasz Domin",
      "email": "lukaszdomin@interia.pl"
    },
    "license": {
      "name": "Licence name",
      "url": "https://some-url.com"
    },
    "version": "0.3"
  },
  "servers": [
    {
      "url": "http://localhost:8081",
      "description": "Local ENV"
    }
  ],
  "security": [
    {
      "bearerAuth": []
    }
  ],
  "paths": {
...
    "/uzytkownicy/send-zmiana-email": {
      "post": {
        "tags": [
          "Uzytkownik"
        ],
        "operationId": "sendChangeEmail",
        "requestBody": {
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/EmailRequest"
              }
            }
          },
          "required": true
        },
        "responses": {
          "200": {
            "description": "OK",
            "content": {
              "application/json": {
                "schema": {
                  "type": "object"
                }
              }
            }
          }
        }
      }
    },
...
    "/uzytkownicy/ustawienia": {
      "get": {
        "tags": [
          "Uzytkownik"
        ],
        "operationId": "getUstawienia",
        "responses": {
          "200": {
            "description": "OK",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/UzytkownikResponse"
                }
              }
            }
          }
        }
      },
      "patch": {
        "tags": [
          "Uzytkownik"
        ],
        "operationId": "updateUstawienia",
        "requestBody": {
          "content": {
            "multipart/form-data": {
              "schema": {
                "required": [
                  "ustawienia"
                ],
                "type": "object",
                "properties": {
                  "ustawienia": {
                    "$ref": "#/components/schemas/UstawieniaRequest"
                  }
                }
              }
            }
          }
        },
        "responses": {
          "200": {
            "description": "OK",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/UzytkownikResponse"
                }
              }
            }
          }
        }
      }
    },
    "/uzytkownicy/profil": {
      "patch": {
        "tags": [
          "Uzytkownik"
        ],
        "operationId": "updateProfil",
        "requestBody": {
          "content": {
            "multipart/form-data": {
              "schema": {
                "required": [
                  "profil"
                ],
                "type": "object",
                "properties": {
                  "profil": {
                    "$ref": "#/components/schemas/ProfilRequest"
                  }
                }
              }
            }
          }
        },
        "responses": {
          "200": {
            "description": "OK",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/UzytkownikResponse"
                }
              }
            }
          }
        }
      }
    },
    "/uzytkownicy/blok/{nazwa}/{blok}": {
      "patch": {
        "tags": [
          "Uzytkownik"
        ],
        "operationId": "setBlokUzytkownik",
        "parameters": [
          {
            "name": "nazwa",
            "in": "path",
            "required": true,
            "schema": {
              "type": "string"
            }
          },
          {
            "name": "blok",
            "in": "path",
            "required": true,
            "schema": {
              "type": "boolean"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "OK",
            "content": {
              "application/json": {
                "schema": {
                  "type": "boolean"
                }
              }
            }
          }
        }
      }
    },
    "/uzytkownicy/avatar": {
      "get": {
        "tags": [
          "Uzytkownik"
        ],
        "operationId": "getAvatar",
        "responses": {
          "200": {
            "description": "OK",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/FileResponse"
                }
              }
            }
          }
        }
      },
      "patch": {
        "tags": [
          "Uzytkownik"
        ],
        "operationId": "updateAvatar",
        "requestBody": {
          "content": {
            "multipart/form-data": {
              "schema": {
                "required": [
                  "file"
                ],
                "type": "object",
                "properties": {
                  "file": {
                    "type": "string",
                    "format": "binary"
                  }
                }
              }
            }
          }
        },
        "responses": {
          "200": {
            "description": "OK",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/UzytkownikResponse"
                }
              }
            }
          }
        }
      }
    },
...
    "/uzytkownicy/szukaj": {
      "get": {
        "tags": [
          "Uzytkownik"
        ],
        "operationId": "findAllUzytkownicy",
        "parameters": [
          {
            "name": "page",
            "in": "query",
            "required": false,
            "schema": {
              "type": "integer",
              "format": "int32",
              "default": 0
            }
          },
          {
            "name": "size",
            "in": "query",
            "required": false,
            "schema": {
              "type": "integer",
              "format": "int32",
              "default": 10
            }
          },
          {
            "name": "szukaj",
            "in": "query",
            "required": false,
            "schema": {
              "type": "string"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "OK",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/PageResponseUzytkownikResponse"
                }
              }
            }
          }
        }
      }
    },
    "/uzytkownicy/profil/{nazwa}": {
      "get": {
        "tags": [
          "Uzytkownik"
        ],
        "operationId": "getStatystykiOfUzytkownik",
        "parameters": [
          {
            "name": "nazwa",
            "in": "path",
            "required": true,
            "schema": {
              "type": "string"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "OK",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/StatystykiDTO"
                }
              }
            }
          }
        }
      }
    },
    "/uzytkownicy/nazwa/{nazwa}": {
      "get": {
        "tags": [
          "Uzytkownik"
        ],
        "operationId": "findByNazwa",
        "parameters": [
          {
            "name": "nazwa",
            "in": "path",
            "required": true,
            "schema": {
              "type": "string"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "OK",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/UzytkownikResponse"
                }
              }
            }
          }
        }
      }
    },
    "/uzytkownicy/email/{email}": {
      "get": {
        "tags": [
          "Uzytkownik"
        ],
        "operationId": "findByEmail",
        "parameters": [
          {
            "name": "email",
            "in": "path",
            "required": true,
            "schema": {
              "type": "string"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "OK",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/UzytkownikResponse"
                }
              }
            }
          }
        }
      }
    },
...
    "/uzytkownicy": {
      "delete": {
        "tags": [
          "Uzytkownik"
        ],
        "operationId": "removeSelf",
        "requestBody": {
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/UsunKontoRequest"
              }
            }
          },
          "required": true
        },
        "responses": {
          "200": {
            "description": "OK"
          }
        }
      }
    },
...
  },
  "components": {
    "schemas": {
...
      "Ustawienia": {
        "type": "object",
        "properties": {
          "id": {
            "type": "integer",
            "format": "int64"
          },
          "statystykiProfilu": {
            "type": "boolean"
          },
          "ogrodPokaz": {
            "type": "boolean"
          },
          "powiadomieniaKomentarzeOdpowiedz": {
            "type": "boolean"
          },
          "powiadomieniaOgrodKwitnienie": {
            "type": "boolean"
          },
          "powiadomieniaOgrodOwocowanie": {
            "type": "boolean"
          }
        }
      },
...
      "EmailRequest": {
        "required": [
          "haslo",
          "nowyEmail"
        ],
        "type": "object",
        "properties": {
          "haslo": {
            "maxLength": 100,
            "minLength": 8,
            "type": "string"
          },
          "nowyEmail": {
            "type": "string"
          }
        }
      },
...
      "UzytkownikResponse": {
        "type": "object",
        "properties": {
          "id": {
            "type": "integer",
            "format": "int64"
          },
          "uzytId": {
            "type": "string"
          },
          "labels": {
            "type": "array",
            "items": {
              "type": "string"
            }
          },
          "nazwa": {
            "type": "string"
          },
          "email": {
            "type": "string"
          },
          "avatar": {
            "type": "string",
            "format": "byte"
          },
          "dataUtworzenia": {
            "type": "string",
            "format": "date-time"
          },
          "aktywowany": {
            "type": "boolean"
          },
          "ban": {
            "type": "boolean"
          },
          "banDo": {
            "type": "string",
            "format": "date"
          },
          "imie": {
            "type": "string"
          },
          "miasto": {
            "type": "string"
          },
          "miejsceZamieszkania": {
            "type": "string"
          },
          "opis": {
            "type": "string"
          },
          "ustawienia": {
            "$ref": "#/components/schemas/Ustawienia"
          },
          "blokowaniUzytkownicy": {
            "uniqueItems": true,
            "type": "array",
            "items": {
              "$ref": "#/components/schemas/UzytkownikResponse"
            }
          },
          "blokujacyUzytkownicy": {
            "uniqueItems": true,
            "type": "array",
            "items": {
              "$ref": "#/components/schemas/UzytkownikResponse"
            }
          }
        }
      },
...
      "UstawieniaRequest": {
        "required": [
          "ogrodPokaz",
          "powiadomieniaKomentarzeOdpowiedz",
          "powiadomieniaOgrodKwitnienie",
          "powiadomieniaOgrodOwocowanie",
          "statystykiProfilu"
        ],
        "type": "object",
        "properties": {
          "statystykiProfilu": {
            "type": "boolean"
          },
          "ogrodPokaz": {
            "type": "boolean"
          },
          "powiadomieniaKomentarzeOdpowiedz": {
            "type": "boolean"
          },
          "powiadomieniaOgrodKwitnienie": {
            "type": "boolean"
          },
          "powiadomieniaOgrodOwocowanie": {
            "type": "boolean"
          }
        }
      },
      "ProfilRequest": {
        "type": "object",
        "properties": {
          "imie": {
            "maxLength": 100,
            "minLength": 0,
            "type": "string"
          },
          "miasto": {
            "maxLength": 100,
            "minLength": 0,
            "type": "string"
          },
          "miejsceZamieszkania": {
            "maxLength": 100,
            "minLength": 0,
            "type": "string"
          },
          "opis": {
            "maxLength": 500,
            "minLength": 0,
            "type": "string"
          }
        }
      },
...
      "FileResponse": {
        "type": "object",
        "properties": {
          "content": {
            "type": "string",
            "format": "byte"
          }
        }
      },
      "PageResponseUzytkownikResponse": {
        "type": "object",
        "properties": {
          "content": {
            "type": "array",
            "items": {
              "$ref": "#/components/schemas/UzytkownikResponse"
            }
          },
          "number": {
            "type": "integer",
            "format": "int32"
          },
          "size": {
            "type": "integer",
            "format": "int32"
          },
          "totalElements": {
            "type": "integer",
            "format": "int64"
          },
          "totalPages": {
            "type": "integer",
            "format": "int32"
          },
          "first": {
            "type": "boolean"
          },
          "last": {
            "type": "boolean"
          }
        }
      },
      "StatystykiDTO": {
        "type": "object",
        "properties": {
          "posty": {
            "type": "integer",
            "format": "int32"
          },
          "komentarze": {
            "type": "integer",
            "format": "int32"
          },
          "rosliny": {
            "type": "integer",
            "format": "int32"
          },
          "komentarzeOcenyPozytywne": {
            "type": "integer",
            "format": "int32"
          },
          "komentarzeOcenyNegatywne": {
            "type": "integer",
            "format": "int32"
          },
          "postyOcenyPozytywne": {
            "type": "integer",
            "format": "int32"
          },
          "postyOcenyNegatywne": {
            "type": "integer",
            "format": "int32"
          }
        }
      },
...
      "UsunKontoRequest": {
        "required": [
          "haslo"
        ],
        "type": "object",
        "properties": {
          "haslo": {
            "maxLength": 100,
            "minLength": 8,
            "type": "string"
          }
        }
      },
...
    },
    "securitySchemes": {
      "bearerAuth": {
        "type": "http",
        "description": "Authorization Bearer JWT",
        "in": "header",
        "scheme": "bearer",
        "bearerFormat": "JWT"
      }
    }
  }
}